+++ src/main/java/org/folio/circulation/support/CqlHelper.java
@@ -1,11 +1,15 @@
 package org.folio.circulation.support;
 
+import io.vertx.core.json.JsonObject;
+import org.joda.time.DateTime;
+import org.joda.time.DateTimeZone;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import java.io.UnsupportedEncodingException;
 import java.lang.invoke.MethodHandles;
 import java.net.URLEncoder;
+import java.nio.charset.StandardCharsets;
 import java.util.Collection;
 import java.util.stream.Collectors;
 
@@ -32,5 +36,25 @@ public class CqlHelper {
         return null;
       }
     }
+  }
+
+  public static String buildisValidUserProxyQuery(JsonObject objectToValidate){
+    //we got the id of the proxy record and user id, look for a record that indicates this is indeed a
+    //proxy for this user id, and make sure that the proxy is valid by indicating that we
+    //only want a match is the status is active and the expDate is in the future
+    String proxyId = objectToValidate.getString("proxyUserId");
+    if(proxyId != null) {
+      DateTime expDate = new DateTime(DateTimeZone.UTC);
+      String validateProxyQuery ="id="+proxyId
+          +" and meta.status=Active"
+          +" and meta.expirationDate>"+expDate.toString().trim();
+      try {
+        return URLEncoder.encode(validateProxyQuery, String.valueOf(StandardCharsets.UTF_8));
+      } catch (UnsupportedEncodingException e) {
+        log.error("Failed to encode query for proxies");
+        return null;
+      }
+    }
+    return null;
   }
 }