+++ src/test/java/org/folio/circulation/api/loans/LoanAPITests.java
@@ -1,10 +1,12 @@
 package org.folio.circulation.api.loans;
 
+import io.vertx.core.json.JsonArray;
 import io.vertx.core.json.JsonObject;
 import org.folio.circulation.api.APITestSuite;
 import org.folio.circulation.api.support.APITests;
 import org.folio.circulation.api.support.builders.ItemRequestBuilder;
 import org.folio.circulation.api.support.builders.LoanRequestBuilder;
+import org.folio.circulation.api.support.builders.UserProxyRequestBuilder;
 import org.folio.circulation.api.support.builders.UserRequestBuilder;
 import org.folio.circulation.api.support.http.ResourceClient;
 import org.folio.circulation.support.JsonArrayHelper;
@@ -31,6 +33,7 @@ import java.util.concurrent.TimeoutException;
 import static org.folio.circulation.api.support.fixtures.UserRequestExamples.basedUponJessicaPontefract;
 import static org.folio.circulation.api.support.fixtures.UserRequestExamples.basedUponStevenJones;
 import static org.folio.circulation.api.support.http.InterfaceUrls.loansUrl;
+import static org.folio.circulation.api.support.http.InterfaceUrls.usersProxyUrl;
 import static org.folio.circulation.api.support.matchers.TextDateTimeMatcher.isEquivalentTo;
 import static org.hamcrest.Matchers.greaterThan;
 import static org.hamcrest.core.Is.is;
@@ -49,9 +52,6 @@ public class LoanAPITests extends APITests {
     UUID id = UUID.randomUUID();
 
     UUID itemId = itemsFixture.basedUponSmallAngryPlanet().getId();
-    String materialTypeId = itemsFixture.basedUponSmallAngryPlanet().getJson().getString("materialTypeId");
-
-    assertThat("materialTypeId is null", materialTypeId == null, is(false));
 
     UUID userId = usersClient.create(new UserRequestBuilder()).getId();
     UUID proxyUserId = UUID.randomUUID();
@@ -62,7 +62,6 @@ public class LoanAPITests extends APITests {
     IndividualResource response = loansClient.create(new LoanRequestBuilder()
       .withId(id)
       .withUserId(userId)
-      .withProxyUserId(proxyUserId)
       .withItemId(itemId)
       .withLoanDate(loanDate)
       .withDueDate(dueDate)
@@ -75,9 +74,6 @@ public class LoanAPITests extends APITests {
 
     assertThat("user id does not match",
       loan.getString("userId"), is(userId.toString()));
-
-    assertThat("proxy user id does not match",
-      loan.getString("proxyUserId"), is(proxyUserId.toString()));
 
     assertThat("item id does not match",
       loan.getString("itemId"), is(itemId.toString()));
@@ -108,15 +104,19 @@ public class LoanAPITests extends APITests {
     assertThat(loan.getJsonObject("item").encode() + " contains 'materialType'",
       loan.getJsonObject("item").containsKey("materialType"), is(true));
 
-    assertThat("materialType exists for item", loan.getJsonObject("item")
-      .containsKey("materialType"), is(true));
-
     assertThat("materialType is book", loan.getJsonObject("item")
       .getJsonObject("materialType").getString("name"), is("Book"));
 
-    assertThat("Joe Smith is a contributor",
-      loan.getJsonObject("item").getJsonArray("contributors")
-        .getJsonObject(0).getString("name"), is("Smith, Joe"));
+    assertThat("item has contributors",
+      loan.getJsonObject("item").containsKey("contributors"), is(true));
+
+    JsonArray contributors = loan.getJsonObject("item").getJsonArray("contributors");
+
+    assertThat("item has a single contributor",
+      contributors.size(), is(1));
+
+    assertThat("Becky Chambers is a contributor",
+      contributors.getJsonObject(0).getString("name"), is("Chambers, Becky"));
 
     assertThat("has item status",
       loan.getJsonObject("item").containsKey("status"), is(true));
@@ -139,10 +139,6 @@ public class LoanAPITests extends APITests {
     assertThat("item status snapshot in storage is not checked out",
       loansStorageClient.getById(id).getJson().getString("itemStatus"),
       is("Checked out"));
-
-
-
-
   }
 
   @Test
@@ -155,21 +151,19 @@ public class LoanAPITests extends APITests {
     UUID id = UUID.randomUUID();
     UUID itemId = itemsFixture.basedUponSmallAngryPlanet().getId();
     UUID userId = usersClient.create(new UserRequestBuilder()).getId();
-    UUID proxyUserId = UUID.randomUUID();
 
     DateTime loanDate = new DateTime(2017, 2, 27, 10, 23, 43, DateTimeZone.UTC);
     DateTime dueDate = new DateTime(2017, 3, 29, 10, 23, 43, DateTimeZone.UTC);
     DateTime systemReturnDate = new DateTime(2017, 4, 1, 12, 0, 0, DateTimeZone.UTC);
 
     JsonObject builtRequest = new LoanRequestBuilder()
+      .closed()
       .withId(id)
       .withUserId(userId)
-      .withProxyUserId(proxyUserId)
       .withItemId(itemId)
       .withLoanDate(loanDate)
       .withDueDate(dueDate)
       .withSystemReturnDate(systemReturnDate)
-      .withStatus("Closed")
       .create();
 
     IndividualResource response = loansClient.create(builtRequest);
@@ -309,7 +303,6 @@ public class LoanAPITests extends APITests {
     loansClient.create(new LoanRequestBuilder()
       .withId(id)
       .withUserId(userId)
-      .withProxyUserId(proxyUserId)
       .withItemId(itemId)
       .withLoanDate(new DateTime(2016, 10, 15, 8, 26, 53, DateTimeZone.UTC))
       .withDueDate(dueDate)
@@ -327,9 +320,6 @@ public class LoanAPITests extends APITests {
 
     assertThat("user id does not match",
       loan.getString("userId"), is(userId.toString()));
-
-    assertThat("proxy user id does not match",
-      loan.getString("proxyUserId"), is(proxyUserId.toString()));
 
     assertThat("item id does not match",
       loan.getString("itemId"), is(itemId.toString()));
@@ -356,6 +346,26 @@ public class LoanAPITests extends APITests {
     assertThat("barcode is taken from item",
       loan.getJsonObject("item").getString("barcode"),
       is("036000291452"));
+
+    assertThat("call number is 123456", loan.getJsonObject("item")
+      .getString("callNumber"), is("123456"));
+
+    assertThat(loan.getJsonObject("item").encode() + " contains 'materialType'",
+      loan.getJsonObject("item").containsKey("materialType"), is(true));
+
+    assertThat("materialType is book", loan.getJsonObject("item")
+      .getJsonObject("materialType").getString("name"), is("Book"));
+
+    assertThat("item has contributors",
+      loan.getJsonObject("item").containsKey("contributors"), is(true));
+
+    JsonArray contributors = loan.getJsonObject("item").getJsonArray("contributors");
+
+    assertThat("item has a single contributor",
+      contributors.size(), is(1));
+
+    assertThat("Becky Chambers is a contributor",
+      contributors.getJsonObject(0).getString("name"), is("Chambers, Becky"));
 
     assertThat("has item status",
       loan.getJsonObject("item").containsKey("status"), is(true));
@@ -979,6 +989,261 @@ public class LoanAPITests extends APITests {
     assertThat(getResponse.getStatusCode(), is(HttpURLConnection.HTTP_NOT_FOUND));
   }
 
+  @Test
+  public void createLoanProxyForTest()
+    throws InterruptedException,
+    ExecutionException,
+    TimeoutException,
+    MalformedURLException {
+
+    UUID id = UUID.randomUUID();
+
+    //create item
+    UUID itemId = itemsFixture.basedUponSmallAngryPlanet().getId();
+
+    //create user
+    UUID userId = usersClient.create(new UserRequestBuilder()).getId();
+
+    //create proxy that is valid with an expDate in the year 2999
+    UUID proxyUserId = UUID.randomUUID();
+    DateTime expDate = new DateTime(2999, 2, 27, 10, 23, 43, DateTimeZone.UTC);
+    UUID recordId = userProxyClient.create(new UserProxyRequestBuilder().
+      withValidationFields(expDate.toString(), "Active", userId.toString(), proxyUserId.toString())).getId();
+
+    //create loan with the proxy id annd user id
+
+    //should pass, do same but create invalid proxy (not active) and check
+    DateTime loanDate = new DateTime(2017, 2, 27, 10, 23, 43, DateTimeZone.UTC);
+    DateTime dueDate = new DateTime(2017, 3, 29, 10, 23, 43, DateTimeZone.UTC);
+
+    //create loan should be allowed as proxy is valid
+    IndividualResource response = loansClient.create(new LoanRequestBuilder()
+      .withId(id)
+      .withUserId(userId)
+      .withProxyUserId(recordId)
+      .withItemId(itemId)
+      .withLoanDate(loanDate)
+      .withDueDate(dueDate)
+      .withStatus("Open"));
+
+    JsonObject loan = response.getJson();
+
+    assertThat("user id does not match",
+      loan.getString("userId"), is(userId.toString()));
+
+    assertThat("proxy user id does not match",
+      loan.getString("proxyUserId"), is(recordId.toString()));
+
+    CompletableFuture<Response> getCompleted = new CompletableFuture<>();
+
+    client.get(usersProxyUrl(""),
+      ResponseHandler.any(getCompleted));
+
+    Response getResponse = getCompleted.get(5, TimeUnit.SECONDS);
+
+    assertThat(getResponse.getStatusCode(), is(HttpURLConnection.HTTP_OK));
+
+    //create proxy that is invalid with an expDate in the year 2000
+    id = UUID.randomUUID();
+
+    proxyUserId = UUID.randomUUID();
+    expDate = new DateTime(2000, 2, 27, 10, 23, 43, DateTimeZone.UTC);
+    recordId = userProxyClient.create(new UserProxyRequestBuilder().
+      withValidationFields(expDate.toString(), "Active", userId.toString(), proxyUserId.toString())).getId();
+
+    loanDate = new DateTime(2017, 2, 27, 10, 23, 43, DateTimeZone.UTC);
+    dueDate = new DateTime(2017, 3, 29, 10, 23, 43, DateTimeZone.UTC);
+
+    //create loan should not be allowed as proxy is valid
+    loan = new LoanRequestBuilder()
+    .withId(id)
+    .withUserId(userId)
+    .withProxyUserId(recordId)
+    .withItemId(itemId)
+    .withLoanDate(loanDate)
+    .withDueDate(dueDate)
+    .withStatus("Open").create();
+
+    CompletableFuture<Response> postCompleted = new CompletableFuture<>();
+
+    client.post(loansUrl(), loan,
+      ResponseHandler.any(postCompleted));
+
+    Response postResponse = postCompleted.get(5, TimeUnit.SECONDS);
+
+    assertThat(postResponse.getBody(), postResponse.getStatusCode(), is(422));
+  }
+
+  @Test
+  public void nonExistentProxy()
+    throws InterruptedException,
+    ExecutionException,
+    TimeoutException,
+    MalformedURLException {
+
+    UUID id = UUID.randomUUID();
+
+    DateTime loanDate = new DateTime(2017, 2, 27, 10, 23, 43, DateTimeZone.UTC);
+    DateTime dueDate = new DateTime(2017, 3, 29, 10, 23, 43, DateTimeZone.UTC);
+
+    JsonObject loan = new LoanRequestBuilder()
+    .withId(id)
+    .withUserId(UUID.randomUUID())
+    .withProxyUserId(UUID.randomUUID())
+    .withItemId(UUID.randomUUID())
+    .withLoanDate(loanDate)
+    .withDueDate(dueDate)
+    .withStatus("Open").create();
+
+    CompletableFuture<Response> postCompleted = new CompletableFuture<>();
+
+    client.post(loansUrl(), loan,
+      ResponseHandler.any(postCompleted));
+
+    Response postResponse = postCompleted.get(5, TimeUnit.SECONDS);
+
+    assertThat(postResponse.getBody(), postResponse.getStatusCode(), is(422));
+
+  }
+
+  @Test
+  public void updateWithValidProxy()
+    throws InterruptedException,
+    ExecutionException,
+    TimeoutException,
+    MalformedURLException {
+
+    UUID id = UUID.randomUUID();
+    UUID itemId = itemsFixture.basedUponSmallAngryPlanet().getId();
+    UUID userId = usersClient.create(new UserRequestBuilder()).getId();
+
+    DateTime loanDate = new DateTime(2017, 2, 27, 10, 23, 43, DateTimeZone.UTC);
+    DateTime dueDate = new DateTime(2017, 3, 29, 10, 23, 43, DateTimeZone.UTC);
+
+    loansClient.create(new LoanRequestBuilder()
+      .withId(id)
+      .withUserId(userId)
+      .withItemId(itemId)
+      .withLoanDate(loanDate)
+      .withDueDate(dueDate)
+      .withStatus("Open"));
+
+    UUID proxyUserId = UUID.randomUUID();
+    DateTime expDate = new DateTime(2999, 2, 27, 10, 23, 43, DateTimeZone.UTC);
+    UUID recordId = userProxyClient.create(new UserProxyRequestBuilder().
+      withValidationFields(expDate.toString(), "Active", userId.toString(), proxyUserId.toString())).getId();
+
+    JsonObject loan = new LoanRequestBuilder()
+    .withId(id)
+    .withUserId(userId)
+    .withProxyUserId(recordId)
+    .withItemId(itemId)
+    .withLoanDate(loanDate)
+    .withDueDate(dueDate)
+    .withStatus("Open").create();
+
+    CompletableFuture<Response> putCompleted = new CompletableFuture<>();
+
+    client.put(loansUrl(String.format("/%s", id)), loan,
+      ResponseHandler.any(putCompleted));
+
+    Response putResponse = putCompleted.get(5, TimeUnit.SECONDS);
+
+    assertThat("Valid proxy should allow updates to work but does not", putResponse.getStatusCode(),
+      is(HttpURLConnection.HTTP_NO_CONTENT));
+
+  }
+
+  @Test
+  public void updateWithInValidProxy()
+    throws InterruptedException,
+    ExecutionException,
+    TimeoutException,
+    MalformedURLException {
+
+    UUID id = UUID.randomUUID();
+    UUID itemId = itemsFixture.basedUponSmallAngryPlanet().getId();
+    UUID userId = usersClient.create(new UserRequestBuilder()).getId();
+
+    DateTime loanDate = new DateTime(2017, 2, 27, 10, 23, 43, DateTimeZone.UTC);
+    DateTime dueDate = new DateTime(2017, 3, 29, 10, 23, 43, DateTimeZone.UTC);
+
+    loansClient.create(new LoanRequestBuilder()
+      .withId(id)
+      .withUserId(userId)
+      .withItemId(itemId)
+      .withLoanDate(loanDate)
+      .withDueDate(dueDate)
+      .withStatus("Open"));
+
+    UUID proxyUserId = UUID.randomUUID();
+    DateTime expDate = new DateTime(1999, 2, 27, 10, 23, 43, DateTimeZone.UTC);
+    UUID recordId = userProxyClient.create(new UserProxyRequestBuilder().
+      withValidationFields(expDate.toString(), "Active", userId.toString(), proxyUserId.toString())).getId();
+
+    JsonObject loan = new LoanRequestBuilder()
+    .withId(id)
+    .withUserId(userId)
+    .withProxyUserId(recordId)
+    .withItemId(itemId)
+    .withLoanDate(loanDate)
+    .withDueDate(dueDate)
+    .withStatus("Open").create();
+
+    CompletableFuture<Response> putCompleted = new CompletableFuture<>();
+
+    client.put(loansUrl(String.format("/%s", id)), loan,
+      ResponseHandler.any(putCompleted));
+
+    Response putResponse = putCompleted.get(5, TimeUnit.SECONDS);
+
+    assertThat("InValid proxy should fail loan update", putResponse.getStatusCode(),
+      is(422));
+
+  }
+
+  @Test
+  public void updateWithNonExistentProxy()
+    throws InterruptedException,
+    ExecutionException,
+    TimeoutException,
+    MalformedURLException {
+
+    UUID id = UUID.randomUUID();
+    UUID itemId = itemsFixture.basedUponSmallAngryPlanet().getId();
+    UUID userId = usersClient.create(new UserRequestBuilder()).getId();
+
+    DateTime loanDate = new DateTime(2017, 2, 27, 10, 23, 43, DateTimeZone.UTC);
+    DateTime dueDate = new DateTime(2017, 3, 29, 10, 23, 43, DateTimeZone.UTC);
+
+    loansClient.create(new LoanRequestBuilder()
+      .withId(id)
+      .withUserId(userId)
+      .withItemId(itemId)
+      .withLoanDate(loanDate)
+      .withDueDate(dueDate)
+      .withStatus("Open"));
+
+    JsonObject loan = new LoanRequestBuilder()
+    .withId(id)
+    .withUserId(userId)
+    .withProxyUserId(UUID.randomUUID())
+    .withItemId(itemId)
+    .withLoanDate(loanDate)
+    .withDueDate(dueDate)
+    .withStatus("Open").create();
+
+    CompletableFuture<Response> putCompleted = new CompletableFuture<>();
+
+    client.put(loansUrl(String.format("/%s", id)), loan,
+      ResponseHandler.any(putCompleted));
+
+    Response putResponse = putCompleted.get(5, TimeUnit.SECONDS);
+
+    assertThat("Invalid proxyUserId is not allowed when updating a loan", putResponse.getStatusCode(), is(422));
+
+  }
+
   private void loanHasExpectedProperties(JsonObject loan) {
     hasProperty("id", loan, "loan");
     hasProperty("userId", loan, "loan");
@@ -994,6 +1259,24 @@ public class LoanAPITests extends APITests {
     hasProperty("barcode", item, "item");
     hasProperty("status", item, "item");
     hasProperty("materialType", item, "item");
+    hasProperty("callNumber", item, "item");
+    hasProperty("contributors", item, "item");
+
+    JsonObject materialType = item.getJsonObject("materialType");
+
+    hasProperty("name", materialType, "material type");
+
+    JsonObject itemStatus = item.getJsonObject("status");
+
+    hasProperty("name", itemStatus, "item status");
+
+    List<JsonObject> contributors = JsonArrayHelper.toList(item.getJsonArray("contributors"));
+
+    assertThat("Should have single contributor",
+      contributors.size(), is(1));
+
+    assertThat("Contributor has a name",
+      contributors.get(0).containsKey("name"), is(true));
 
     assertThat("Should not have snapshot of item status, as current status is included",
       loan.containsKey("itemStatus"), is(false));