+++ src/test/java/org/folio/circulation/api/requests/RequestsAPIUpdatingTests.java
@@ -1,18 +1,9 @@
 package org.folio.circulation.api.requests;
 
-import io.vertx.core.json.JsonObject;
-import org.folio.circulation.api.support.APITests;
-import org.folio.circulation.api.support.builders.ItemRequestBuilder;
-import org.folio.circulation.api.support.builders.RequestRequestBuilder;
-import org.folio.circulation.api.support.builders.UserRequestBuilder;
-import org.folio.circulation.api.support.http.InterfaceUrls;
-import org.folio.circulation.support.http.client.IndividualResource;
-import org.folio.circulation.support.http.client.Response;
-import org.folio.circulation.support.http.client.ResponseHandler;
-import org.joda.time.DateTime;
-import org.joda.time.DateTimeZone;
-import org.joda.time.LocalDate;
-import org.junit.Test;
+import static org.folio.circulation.api.support.fixtures.LoanFixture.checkOutItem;
+import static org.folio.circulation.api.support.matchers.TextDateTimeMatcher.isEquivalentTo;
+import static org.hamcrest.core.Is.is;
+import static org.hamcrest.junit.MatcherAssert.assertThat;
 
 import java.io.UnsupportedEncodingException;
 import java.net.HttpURLConnection;
@@ -23,10 +14,21 @@ import java.util.concurrent.ExecutionException;
 import java.util.concurrent.TimeUnit;
 import java.util.concurrent.TimeoutException;
 
-import static org.folio.circulation.api.support.fixtures.LoanFixture.checkOutItem;
-import static org.folio.circulation.api.support.matchers.TextDateTimeMatcher.isEquivalentTo;
-import static org.hamcrest.core.Is.is;
-import static org.hamcrest.junit.MatcherAssert.assertThat;
+import org.folio.circulation.api.support.APITests;
+import org.folio.circulation.api.support.builders.ItemRequestBuilder;
+import org.folio.circulation.api.support.builders.RequestRequestBuilder;
+import org.folio.circulation.api.support.builders.UserProxyRequestBuilder;
+import org.folio.circulation.api.support.builders.UserRequestBuilder;
+import org.folio.circulation.api.support.http.InterfaceUrls;
+import org.folio.circulation.support.http.client.IndividualResource;
+import org.folio.circulation.support.http.client.Response;
+import org.folio.circulation.support.http.client.ResponseHandler;
+import org.joda.time.DateTime;
+import org.joda.time.DateTimeZone;
+import org.joda.time.LocalDate;
+import org.junit.Test;
+
+import io.vertx.core.json.JsonObject;
 
 public class RequestsAPIUpdatingTests extends APITests {
   @Test
@@ -520,5 +522,101 @@ public class RequestsAPIUpdatingTests extends APITests {
     assertThat("barcode is not taken from item",
       representation.getJsonObject("item").containsKey("barcode"),
       is(false));
+  }
+
+  @Test
+  public void canUpdateARequestForItemWithValidUserProxy()
+    throws InterruptedException,
+    ExecutionException,
+    TimeoutException,
+    MalformedURLException {
+
+    UUID id = UUID.randomUUID();
+
+    UUID itemId = itemsFixture.basedUponTemeraire().getId();
+
+    checkOutItem(itemId, loansClient);
+
+    UUID requesterId = usersClient.create(new UserRequestBuilder()).getId();
+
+    DateTime requestDate = new DateTime(2017, 7, 22, 10, 22, 54, DateTimeZone.UTC);
+
+    IndividualResource createdRequest = requestsClient.create(
+      new RequestRequestBuilder()
+      .recall()
+      .withId(id)
+      .withRequestDate(requestDate)
+      .withItemId(itemId)
+      .withRequesterId(requesterId)
+      .fulfilToHoldShelf()
+      .withRequestExpiration(new LocalDate(2017, 7, 30))
+      .withHoldShelfExpiration(new LocalDate(2017, 8, 31)));
+
+    DateTime expDate = new DateTime(2999, 2, 27, 10, 23, 43, DateTimeZone.UTC);
+    UUID recordId = userProxyClient.create(new UserProxyRequestBuilder().
+      withValidationFields(expDate.toString(), "Active",
+        UUID.randomUUID().toString(), UUID.randomUUID().toString())).getId();
+
+    JsonObject updatedRequest = createdRequest.copyJson();
+
+    updatedRequest
+      .put("proxyUserId", recordId.toString());
+
+    CompletableFuture<Response> putCompleted = new CompletableFuture<>();
+
+    client.put(InterfaceUrls.requestsUrl(String.format("/%s", id)),
+      updatedRequest, ResponseHandler.any(putCompleted));
+
+    Response putResponse = putCompleted.get(5, TimeUnit.SECONDS);
+
+    assertThat(putResponse.getStatusCode(), is(HttpURLConnection.HTTP_NO_CONTENT));
+  }
+
+  @Test
+  public void canNotUpdateARequestForItemWithInValidUserProxy()
+    throws InterruptedException,
+    ExecutionException,
+    TimeoutException,
+    MalformedURLException {
+
+    UUID id = UUID.randomUUID();
+
+    UUID itemId = itemsFixture.basedUponTemeraire().getId();
+
+    checkOutItem(itemId, loansClient);
+
+    UUID requesterId = usersClient.create(new UserRequestBuilder()).getId();
+
+    DateTime requestDate = new DateTime(2017, 7, 22, 10, 22, 54, DateTimeZone.UTC);
+
+    IndividualResource createdRequest = requestsClient.create(
+      new RequestRequestBuilder()
+      .recall()
+      .withId(id)
+      .withRequestDate(requestDate)
+      .withItemId(itemId)
+      .withRequesterId(requesterId)
+      .fulfilToHoldShelf()
+      .withRequestExpiration(new LocalDate(2017, 7, 30))
+      .withHoldShelfExpiration(new LocalDate(2017, 8, 31)));
+
+    DateTime expDate = new DateTime(1999, 2, 27, 10, 23, 43, DateTimeZone.UTC);
+    UUID recordId = userProxyClient.create(new UserProxyRequestBuilder().
+      withValidationFields(expDate.toString(), "Active",
+        UUID.randomUUID().toString(), UUID.randomUUID().toString())).getId();
+
+    JsonObject updatedRequest = createdRequest.copyJson();
+
+    updatedRequest
+      .put("proxyUserId", recordId.toString());
+
+    CompletableFuture<Response> putCompleted = new CompletableFuture<>();
+
+    client.put(InterfaceUrls.requestsUrl(String.format("/%s", id)),
+      updatedRequest, ResponseHandler.any(putCompleted));
+
+    Response putResponse = putCompleted.get(5, TimeUnit.SECONDS);
+
+    assertThat(putResponse.getStatusCode(), is(422));
   }
 }