+++ src/test/java/org/folio/circulation/api/support/fixtures/InstanceRequestExamples.java
@@ -4,27 +4,33 @@ import org.folio.circulation.api.support.builders.InstanceRequestBuilder;
 
 public class InstanceRequestExamples {
   public static InstanceRequestBuilder basedUponSmallAngryPlanet() {
-    return create("The Long Way to a Small, Angry Planet");
+    return create("The Long Way to a Small, Angry Planet")
+      .withContributor("Chambers, Becky");
   }
 
   public static InstanceRequestBuilder basedUponNod() {
-    return create("Nod");
+    return create("Nod")
+      .withContributor("Barnes, Adrian");
   }
 
   public static InstanceRequestBuilder basedUponUprooted() {
-    return create("Uprooted");
+    return create("Uprooted")
+      .withContributor("Novik, Naomi");
   }
 
   public static InstanceRequestBuilder basedUponTemeraire() {
-    return create("Temeraire");
+    return create("Temeraire")
+      .withContributor("Novik, Naomi");
   }
 
   public static InstanceRequestBuilder basedUponInterestingTimes() {
-    return create("Interesting Times");
+    return create("Interesting Times")
+      .withContributor("Pratchett, Terry");
   }
 
   public static InstanceRequestBuilder basedUponDunkirk() {
-    return new InstanceRequestBuilder("Dunkirk");
+    return new InstanceRequestBuilder("Dunkirk")
+      .withContributor("Nolan, Christopher");
   }
 
   private static InstanceRequestBuilder create(String title) {